pyproject.toml
[project]
name = "lavendar-ai-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "groq>=0.31.0",
    "opencv-python>=4.12.0.88",
    "python-dotenv>=1.0.1",
    "langchain-google-genai",
    "langgraph",
    "speechrecognition>=3.14.3",
    "pydub>=0.25.1",
    "pyaudio>=0.2.14",
]
